Fixes for both HourBoostr and SingleBoostr
Added invoke call to UI when game runs out of cards.
Fixed some incorrect strings in both Single and HourBoostr. Upgraded
HourBoostr SteamKit2 and Newtonsoft.Json packages. Added blacklisting to
SingleBoostr.

+++ SingleBoostr/SingleBoostr/SingleBoostr/NativeMethods.cs
@@ -1,8 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
 using System.Runtime.InteropServices;
 
 namespace SingleBoostr
@@ -10,7 +7,7 @@ namespace SingleBoostr
     class NativeMethods
     {
         [DllImport("user32.dll")]
-        public static extern int SendMessage(IntPtr hWnd, int Msg, int wParam, int lParam);
+        public static extern IntPtr SendMessage(IntPtr hWnd, UInt32 Msg, Int32 wParam, Int32 lParam);
 
         [DllImport("user32.dll", CharSet = CharSet.Auto)]
         public static extern int SendMessage(IntPtr hWnd, uint Msg, IntPtr wParam, [MarshalAs(UnmanagedType.LPWStr)]string lParam);
